+++ ngDesk-Company-Service/src/main/java/com/ngdesk/company/CompanyServiceApplication.java
@@ -5,8 +5,10 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.ComponentScan;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
 @EnableDiscoveryClient
+@EnableScheduling
 @EnableFeignClients({"com.ngdesk.commons", "com.ngdesk.company"})
 @ComponentScan({"com.ngdesk.commons", "com.ngdesk.company", "com.ngdesk.repositories", "com.ngdesk.workflow.dao"})
 @SpringBootApplication
