+++ ngDesk-Company-Service/src/main/resources/daily_company_signup.html
@@ -1,0 +1,7 @@
+<!DOCTYPE html>
+<head></head>
+<body>
+   List of companies signed up for ngDesk.<br><br>
+   TABLE_REPLACE
+   <body>
+</html>
